requirements.txt
requests==2.31.0
boto3==1.28.53
Pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
tqdm==4.66.1